# perf(propagate): don't increase the stack when there is only one subscriber