@johnsoncodehk 2024-11-27
Before this PR, propagate always used stack to fall back to the previous node.

<img width="250" alt="image" src="https://github.com/user-attachments/assets/17ed322e-9a46-4ca0-b535-d1eaff03550b">

In this PR, propagate will only roll back to the previous branch.

<img width="280" alt="image" src="https://github.com/user-attachments/assets/6130cfed-8335-4747-8a44-e4e7cb7cfab5">